Rainbow Sword Memory allocation
Changed Rainbow Swords to take up a byte instead of a double, and hopefully fixing patch file generation with this change.

+++ ASM/c/rainbow_sword.c
@@ -9,7 +9,7 @@ uint32_t frames = 0;
 
 const uint32_t CYCLE_FRAMES = 0x30;
 
-extern uint32_t RAINBOW_SWORD_ENABLED;
+extern uint8_t RAINBOW_SWORD_ENABLED;
 
 typedef struct
 {
